+++ doppler/src/oracle.rs
@@ -10,7 +10,7 @@ pub struct Oracle<T: Sized + Copy> {
 
 impl<T: Sized + Copy> Oracle<T> {
     // Relative offsets for instruction data
-    const INSTRUCTION_SEQUENCE: usize = 0x50d8 + core::mem::size_of::<T>(); // (sequence: u64) 
+    const INSTRUCTION_SEQUENCE: usize = 0x50d8 + core::mem::size_of::<T>(); // (sequence: u64)
     const INSTRUCTION_PAYLOAD: usize = 0x50e0 + core::mem::size_of::<T>(); // (payload: T)
 
     /// # Safety
