+++ sdk/Cargo.toml
@@ -14,4 +14,7 @@ solana-account = { workspace = true }
 solana-compute-budget-interface = { workspace = true }
 solana-instruction = { workspace = true }
 solana-pubkey = { workspace = true }
+solana-keypair = { workspace = true }
+solana-hash = { workspace = true }
+solana-signer = { workspace = true }
 solana-transaction = { workspace = true }
