+++ sdk/src/lib.rs
@@ -1,292 +1,5 @@
-use solana_compute_budget_interface::ComputeBudgetInstruction;
-use solana_instruction::{AccountMeta, Instruction};
-use solana_pubkey::Pubkey;
-use solana_transaction::Transaction;
-
-// fastRQJt3nLdY3QA7n8eZ8ETEVefy56ryfUGVkfZokm
-pub const ID: Pubkey = Pubkey::new_from_array([
-    0x09, 0xe2, 0x60, 0x40, 0xff, 0x10, 0xec, 0xcf, 0xc1, 0x6a, 0xf6, 0x16, 0x9a, 0x68, 0x04, 0x78,
-    0x15, 0x14, 0x33, 0x02, 0xac, 0x6e, 0x98, 0x5f, 0x70, 0x85, 0x53, 0xe1, 0x0a, 0xb6, 0xf9, 0x22,
-]);
-
-const SEQUENCE_CHECK_CU: u32 = 5;
-const ADMIN_VERIFICATION_CU: u32 = 6;
-const PAYLOAD_WRITE_CU: u32 = 6;
-
-const COMPUTE_BUDGET_IXS_NUM: u32 = 3;
-const COMPUTE_BUDGET_IXS_CU_OVERHEAD: u32 = COMPUTE_BUDGET_IXS_NUM * 150; // 3 compute budget ixs * 150 CU each
-const COMPUTE_BUDGET_UNIT_PRICE_SIZE: u32 = 9;
-const COMPUTE_BUDGET_UNIT_LIMIT_SIZE: u32 = 5;
-const COMPUTE_BUDGET_DATA_LIMIT_SIZE: u32 = 5;
-const COMPUTE_BUDGET_PROGRAM_SIZE: u32 = 22;
-const ORACLE_PROGRAM_SIZE: u32 = 36;
-const DATA_SIZE_OVERHEAD: u32 = ORACLE_PROGRAM_SIZE
-    + COMPUTE_BUDGET_PROGRAM_SIZE
-    + COMPUTE_BUDGET_UNIT_PRICE_SIZE
-    + COMPUTE_BUDGET_UNIT_LIMIT_SIZE
-    + COMPUTE_BUDGET_DATA_LIMIT_SIZE
-    + 2;
-
-pub struct TransactionBuilder {
-    oracle_update_ixs: Vec<Instruction>,
-    admin: Pubkey,
-    unit_price: u64,
-    compute_units: u32,
-    loaded_account_data_size: u32,
-}
-
-impl TransactionBuilder {
-    pub fn new(admin: Pubkey) -> Self {
-        Self {
-            admin,
-            oracle_update_ixs: vec![],
-            unit_price: 1,
-            compute_units: 0,
-            loaded_account_data_size: 0,
-        }
-    }
-
-    pub fn add_oracle_update_instruction<T: Sized + Copy>(
-        mut self,
-        oracle_pubkey: Pubkey,
-        oracle: Oracle<T>,
-    ) -> Self {
-        let update_ix = UpdateInstruction {
-            admin: self.admin,
-            oracle_pubkey,
-            oracle,
-        };
-
-        self.compute_units += update_ix.compute_units();
-        self.loaded_account_data_size += update_ix.loaded_accounts_data_size_limit() * 2;
-
-        self.oracle_update_ixs.push(update_ix.into());
-
-        self
-    }
-
-    pub fn unit_price(mut self, micro_lamports: u64) -> Self {
-        self.unit_price = micro_lamports;
-        self
-    }
-
-    pub fn build(self) -> Vec<Instruction> {
-        let mut ixs =
-            Vec::with_capacity(self.oracle_update_ixs.len() + COMPUTE_BUDGET_IXS_NUM as usize);
-
-        ixs.push(ComputeBudgetInstruction::set_compute_unit_price(
-            self.unit_price,
-        ));
-        ixs.push(ComputeBudgetInstruction::set_compute_unit_limit(
-            self.compute_units + COMPUTE_BUDGET_IXS_CU_OVERHEAD,
-        ));
-        ixs.push(
-            ComputeBudgetInstruction::set_loaded_accounts_data_size_limit(
-                self.loaded_account_data_size + DATA_SIZE_OVERHEAD,
-            ),
-        );
-
-        for oracle_ix in self.oracle_update_ixs {
-            ixs.push(oracle_ix);
-        }
-
-        ixs
-    }
-}
-
-#[repr(C)]
-#[derive(Clone, Copy, Debug)]
-pub struct Oracle<T: Sized + Copy> {
-    pub sequence: u64,
-    pub payload: T,
-}
-
-impl<T: Sized + Copy> Oracle<T> {
-    pub fn to_bytes(&self) -> Vec<u8> {
-        let mut data = Vec::with_capacity(core::mem::size_of::<Self>());
-        // write sequence bytes
-        data.extend_from_slice(&self.sequence.to_le_bytes());
-        // write payload bytes
-        data.extend_from_slice(unsafe {
-            core::slice::from_raw_parts(
-                core::ptr::from_ref(&self.payload) as *const u8,
-                core::mem::size_of::<T>(),
-            )
-        });
-        data
-    }
-
-    pub fn from_bytes(data: &[u8]) -> Self {
-        assert!(data.len() == core::mem::size_of::<Self>());
-
-        // read u64 sequence from first 8 bytes
-        let mut seq_bytes = [0u8; 8];
-        seq_bytes.copy_from_slice(&data[..8]);
-        let sequence = u64::from_le_bytes(seq_bytes);
-
-        // read payload from remaining bytes
-        let payload = unsafe { *(data[8..].as_ptr() as *const T) };
-
-        Self { sequence, payload }
-    }
-}
-
-pub struct UpdateInstruction<T: Sized + Copy> {
-    pub admin: Pubkey,
-    pub oracle_pubkey: Pubkey,
-    pub oracle: Oracle<T>,
-}
-
-impl<T: Sized + Copy> UpdateInstruction<T> {
-    pub const fn compute_units(&self) -> u32 {
-        SEQUENCE_CHECK_CU
-            + ADMIN_VERIFICATION_CU
-            + PAYLOAD_WRITE_CU
-            + (core::mem::size_of::<Oracle<T>>() / 4) as u32
-    }
-
-    pub const fn loaded_accounts_data_size_limit(&self) -> u32 {
-        core::mem::size_of::<Oracle<T>>() as u32
-    }
-}
-
-impl<T: Sized + Copy> From<UpdateInstruction<T>> for Instruction {
-    fn from(update: UpdateInstruction<T>) -> Self {
-        let data = update.oracle.to_bytes();
-
-        Self {
-            program_id: ID,
-            accounts: vec![
-                AccountMeta::new_readonly(update.admin, true),
-                AccountMeta::new(update.oracle_pubkey, false),
-            ],
-            data,
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use doppler_program::PriceFeed;
-
-    use super::*;
-
-    #[repr(C)]
-    #[derive(Clone, Copy)]
-    struct PropAMM {
-        pub bid: u64,
-        pub ask: u64,
-    }
-
-    #[repr(C)]
-    #[derive(Clone, Copy)]
-    struct MarketData {
-        pub price: u64,
-        pub volume: u64,
-        pub confidence: u32,
-    }
-
-    #[test]
-    fn test_oracle_to_bytes() {
-        let oracle = Oracle {
-            sequence: 42,
-            payload: 123u32,
-        };
-
-        let bytes = oracle.to_bytes();
-        assert_eq!(bytes.len(), 12);
-        assert_eq!(&bytes[0..8], &42u64.to_le_bytes());
-        assert_eq!(&bytes[8..12], &123u32.to_le_bytes());
-    }
-
-    #[test]
-    fn test_cu_limit_num_payload() {
-        let admin = Pubkey::new_unique();
-        let oracle_pubkey = Pubkey::new_unique();
-
-        let oracle = Oracle {
-            sequence: 1,
-            payload: 789u64,
-        };
-
-        let update_instruction = UpdateInstruction {
-            admin,
-            oracle_pubkey,
-            oracle,
-        };
-
-        let compute_instruction = update_instruction.compute_units();
-
-        assert_eq!(compute_instruction, 21);
-    }
-
-    #[test]
-    fn test_cu_limit_price_feed_payload() {
-        let admin = Pubkey::new_unique();
-        let oracle_pubkey = Pubkey::new_unique();
-
-        let oracle = Oracle {
-            sequence: 1,
-            payload: PriceFeed { price: 1_100_000 },
-        };
-
-        let update_instruction = UpdateInstruction {
-            admin,
-            oracle_pubkey,
-            oracle,
-        };
-
-        let compute_instruction = update_instruction.compute_units();
-
-        assert_eq!(compute_instruction, 21);
-    }
-
-    #[test]
-    fn test_cu_limit_prop_amm_payload() {
-        let admin = Pubkey::new_unique();
-        let oracle_pubkey = Pubkey::new_unique();
-
-        let oracle = Oracle {
-            sequence: 1,
-            payload: PropAMM {
-                bid: 10_500_000,
-                ask: 10_550_000,
-            },
-        };
-
-        let update_instruction = UpdateInstruction {
-            admin,
-            oracle_pubkey,
-            oracle,
-        };
-
-        let compute_instruction = update_instruction.compute_units();
-
-        assert_eq!(compute_instruction, 23);
-    }
-
-    #[test]
-    fn test_cu_limit_market_data_payload() {
-        let admin = Pubkey::new_unique();
-        let oracle_pubkey = Pubkey::new_unique();
-
-        let oracle = Oracle {
-            sequence: 1,
-            payload: MarketData {
-                price: 45_000_000,
-                volume: 150_000_000,
-                confidence: 300,
-            },
-        };
-
-        let update_instruction = UpdateInstruction {
-            admin,
-            oracle_pubkey,
-            oracle,
-        };
-
-        let compute_instruction = update_instruction.compute_units();
-
-        assert_eq!(compute_instruction, 25);
-    }
-}
+mod accounts;
+mod constants;
+pub mod transaction;
+pub use accounts::{Oracle, UpdateInstruction};
+pub use constants::ID;
